requirements.txt
dataclasses-json==0.5.2
numpy==1.23.4
requests==2.28.1
websocket-client==0.57.0
aiohttp==3.8.3